remove AuctionHistory and data transformation from request and bring them down the client
